Decode 'htmlsrc' to string before re.sub()
Python 3 has distinct types for a byte array and a string, and requires
explicit conversion between them. In order to make the script to work with
Python 3, decode 'htmlsrc' to string before applying using the re.sub() API.
Before writing the attachment to a file, encode it back to as a byte array.

Signed-off-by: Jarkko Sakkinen <jarkko.sakkinen@iki.fi>