New script pip_upgrade.py.
Upgrade every package in the current virtualenv.

Sprung out of discussions like
https://coderwall.com/p/quwaxa/update-all-installed-python-packages-with-pip
and a discussion on #python that revealed that there's really no
sensible way to do that, because the official Python way of working
seems to be to have a separate venv for every program you run, with
versions all pinned to the last time the package was updated.