viewmailattachments: allow optional image viewer.
If IMAGE_VIEWER is set to a command name in the script,
then if the image has top-level images and no multipart/related
parts, use an image viewer to display the images, passing all
the temp file image names at once.

This was inspired by someone who sent me a message with a bunch
of image/heif attachments that qtwebengine couldn't display,
but pho can display them just fine.