eggtimer: add a way to change a running timer's time.
Add a socket protocol that allows for querying current status
(so when querying eggtimers, the results are printed on the current
tty, not the tty from which each eggtimer was started)
and also allows for adding or subtracting minutes.

And yes, this is way overkill for a silly little eggtimer program;
but it's also a good example of how to use unix-domain sockets for
full duplex interprocess communication.