Lots of work on transparent/clickthrough images.
The old transimageviewer.py is obsolete because GTK2.

I've gotten imageviewer3.py working with transparency,
but not with clickthrough.

I tried a Tk approach, transimageviewer-tk.py: it also handles
transparency but not clickthrough. There's some Xlib code there
that should do it (it does it for pure Xlib windows) but it
doesn't work on the Tk window.

Finally, with qt5 I was able to do both transparency and clickthrough,
though not in the obvious way: WA_TransparentForMouseEvents doesn't
do it, you have to make the window completely invisible to the window
manager with X11BypassWindowManagerHint. transimageviewer-qt5.py.

Finally, there's a proof of concept of an Xlib image viewer,
ximageviewer.py. The Xlib shapewin.py demo gives an example
of setting a window's boundary shape, so I have hopes that setting
an input shape (as I tried to do in the Tk version) will work,
though I haven't gotten it working yet.