transimageviewer_x: finish the clickthrough code.
With the input shape enabled, dragging on the titlebar
no longer works.
Instead, I tried to make a band of 20 pixels height across the top
that you can use with a windowmanager's move-window key.
It works inconsistently, and only near the (horizontal)
middle of the top band, not on the corners. Not sure why yet,
but at least it's usable now.