quickbrowse WIP: more pointless flailing on QT6.
No matter what I try I can't seem to get the Qt6 QWebEngineView
to obey a load(url) call when it's in a non-primary tab.
I'm thinking it might be better to switch to qutebrowser for my
lightweight browsing needs, let quickbrowse die when Qt5 ceases
to be maintained, and stop wasting so much of my time trying
to figure out through trial and error what poorly documented Qt
classes might or might not want.