constituents: PickPoint, plus changes to run on the real data.
I have the actual member database now, and have updated
constituents.py accordingly so that it pulls addresses from
the member database and writes to json and csv.

Also included is code to work with the PickPoint API, since
Nominatim randomly refuses to work even if you obey the rate limiter.